refactor(visualization): 可視化関連コードを crane_visualization_interfaces に統合
## 背景
SVG可視化関連のコードが3つのパッケージ（crane_visualization_interfaces・
crane_msg_wrappers・crane_visualization_aggregator）に分散しており、
依存関係や参照先が複雑になっていた。

## 変更内容
- crane_visualization_interfaces にビジュアライザビルダーライブラリと
  アグリゲータノードを追加し、可視化の単一パッケージに昇格
  - CMakeLists.txt をrosidl＋C++ライブラリ＋コンポーネントのハイブリッド構成へ書き換え
  - package.xml に crane_geometry・rclcpp・rclcpp_components 依存を追加
  - crane_visualizer_wrapper.hpp/.cpp を crane_msg_wrappers から移動
  - crane_visualization_aggregator ノードを rclcpp コンポーネント化して移動
- crane_msg_wrappers の crane_visualizer_wrapper.hpp をフォワーディングヘッダに置換
- crane_visualization_aggregator パッケージを削除
- 全コンシューマ（17ファイル）の include パスを
  crane_msg_wrappers → crane_visualization_interfaces に更新
- crane_bringup/launch/crane.launch.xml のアグリゲータノード参照を更新
- docs/visualizer.md のパッケージ・include パス参照を更新
