+++ crane_bringup/launch/crane.launch.xml
@@ -202,7 +202,7 @@ limitations under the License.
   <node pkg="crane_web_debugger" exec="crane_websocket_server" output="screen"/>
 
   <!-- 可視化情報の集約 -->
-  <node pkg="crane_visualization_aggregator" exec="crane_visualization_aggregator_node" output="screen"/>
+  <node pkg="crane_visualization_interfaces" exec="visualization_aggregator_node" output="screen"/>
 
   <!-- ワールドモデル（世界状態の推定と管理） -->
   <node pkg="crane_world_model_publisher" exec="crane_world_model_publisher_node" output="screen">
