+++ crane_game_analyzer/include/crane_game_analyzer/metrics/metric_base.hpp
@@ -7,7 +7,7 @@
 #ifndef CRANE_GAME_ANALYZER__METRICS__METRIC_BASE_HPP_
 #define CRANE_GAME_ANALYZER__METRICS__METRIC_BASE_HPP_
 
-#include <crane_msg_wrappers/crane_visualizer_wrapper.hpp>
+#include <crane_visualization_interfaces/crane_visualizer_wrapper.hpp>
 #include <memory>
 #include <string>
 #include <vector>
