--- crane_visualization_aggregator/package.xml
@@ -1,23 +1,0 @@
-<?xml version="1.0"?>
-<?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
-<package format="3">
-  <name>crane_visualization_aggregator</name>
-  <version>1.0.314</version>
-  <description>crane_visualization_aggregator</description>
-  <maintainer email="ibis.ssl.team@gmail.com">ibis ssl</maintainer>
-  <license>MIT</license>
-
-  <buildtool_depend>ament_cmake_auto</buildtool_depend>
-
-  <depend>crane_msgs</depend>
-  <depend>crane_visualization_interfaces</depend>
-  <depend>rclcpp</depend>
-  <depend>std_msgs</depend>
-
-  <test_depend>ament_lint_auto</test_depend>
-  <test_depend>crane_lint_common</test_depend>
-
-  <export>
-    <build_type>ament_cmake</build_type>
-  </export>
-</package>
