+++ crane_visualization_interfaces/package.xml
@@ -3,7 +3,7 @@
 <package format="3">
   <name>crane_visualization_interfaces</name>
   <version>1.0.314</version>
-  <description>Message types for CON-SAI visualization</description>
+  <description>Message types and visualization builder/aggregator for crane SVG visualization</description>
   <maintainer email="ibis.ssl.team@gmail.com">ibis-ssl</maintainer>
   <license>MIT</license>
 
@@ -13,6 +13,9 @@
   <build_depend>rosidl_default_generators</build_depend>
 
   <depend>builtin_interfaces</depend>
+  <depend>crane_geometry</depend>
+  <depend>rclcpp</depend>
+  <depend>rclcpp_components</depend>
   <depend>std_msgs</depend>
 
   <exec_depend>rosidl_default_runtime</exec_depend>
